backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Face Recognition
opencv-python==4.9.0.80
mtcnn==0.1.1
insightface==0.7.3
onnxruntime==1.17.0
scikit-learn==1.4.0

# Optional: FAISS for fast vector search
faiss-cpu==1.7.4

# Image Processing
Pillow==10.2.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Date/Time
python-dateutil==2.8.2

# Logging
loguru==0.7.2
